+++ YumSale/YumSale.csproj
@@ -184,12 +184,16 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\201503050321121_InitialCreate.cs" />
+    <Compile Include="Migrations\201503050321121_InitialCreate.Designer.cs">
+      <DependentUpon>201503050321121_InitialCreate.cs</DependentUpon>
+    </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
     <Compile Include="Models\Buyer.cs" />
+    <Compile Include="Models\Class1.cs" />
     <Compile Include="Models\Item.cs" />
     <Compile Include="Models\IdentityModels.cs" />
-    <Compile Include="Models\ItemContext.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
@@ -273,6 +277,11 @@
   <ItemGroup>
     <Content Include="packages.config" />
     <None Include="Project_Readme.html" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Migrations\201503050321121_InitialCreate.resx">
+      <DependentUpon>201503050321121_InitialCreate.cs</DependentUpon>
+    </EmbeddedResource>
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
